Migrating grids and levels to Spatial
Adding versioning from both classes in Geometry as well as in Architecture
